requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
keyboard>=0.13.5
boto3>=1.26.0
requests>=2.28.0
